+++ ExtraCrestSlots/changelog.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## [1.0.1]
+- added option 'IsLocked' to have the slot locked (does not add more unlock medallions); not recommended
+- added option 'OverrideSlot' to override color and binding of existing slots
+
 ## [1.0.0]
 - added 6 slots below each crest (first red, if crest has no down activatable)